fix(build): ensure all xgboost submodules are checkouted
Currently dmlc/rabi have been removed from master branch of xgboost

`git clone --recursive` will clone submodule of master branch, so `rabi`
is missing. Then we switch to an old xgboost version and it can't build
because of the missing deps.

This change retrieves submodule after we switch to the wanted commit-id.