refactor: hide visitor complexity inside the visitor itself
This aims to make the code flow looks like we don't have visitor in the
loop.

The idea is to put all visitor of a particular base class, into
dedicated namespace. Each visitor will have a static method in this
namespace where we move all of visitor boilerplate setup. This avoid
creating intermediate object to store the method result. Using a visitor
become a simple as:

  `int status = visitor_mllib::predict_job(mllib, ad_in, ad_out, chain);`

The visitor helper signature looks like the real mllib method.