feat(docker): rework Dockerfile
Except the base image and env DEEPDETECT_ARCH=, the build is the same for cpu and gpu

New docker features are used to speedup the build. (packages and ccache
directory are cached by docker)

The local directory is used to build the image, no need to redownload
the whole deepdetect code. And this allow to build image from other git
reference not only master.